refacto(users): extract organization table data layer
